chart/requirements.txt
boto3==1.16.39
botocore==1.19.39
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
git+ssh://git@github.com/sendbird/sendbird_common.git@v0.0.30#egg=sendbird_common
